module names renaming -- added underlines for specified internal use
