+++ notiblocks/main/_nb_handler.py
@@ -1,17 +1,17 @@
-from .nb_config import NBConfig
-from .enums.brackets import Brackets
+from ._nb_config import NBConfig
+from .enums._brackets import Brackets
 
-from .error.invalid_format_error import InvalidFormatError
-from .error.invalid_operation_type_error import InvalidOperationTypeError
+from .error._invalid_format_error import InvalidFormatError
+from .error._invalid_operation_type_error import InvalidOperationTypeError
 
-from .enums.colors.bgcolors import BGColors
-from .enums.colors.fgcolors import FGColors
+from .enums.colors._bgcolors import BGColors
+from .enums.colors._fgcolors import FGColors
 
-from .constants import RESET_STYLE
+from ._constants import RESET_STYLE
 
-from .enums.timestampformat import TimeStampFormats
+from .enums._timestampformat import TimeStampFormats
 
-from .ansi import ANSI
+from ._ansi import ANSI
 
 from pathlib import Path
