+++ notiblocks/main/enums/colors/_bgcolors.py
@@ -1,6 +1,6 @@
 from enum import Enum
 
-from ...constants import *
+from ..._constants import *
 
 class BGColors(Enum):
     """
