+++ notiblocks/main/enums/colors/_fgcolors.py
@@ -1,6 +1,6 @@
 from enum import Enum
 
-from ...constants import *
+from ..._constants import *
 
 class FGColors(Enum):
     """
