+++ notiblocks/notiblocks.py
@@ -1,6 +1,9 @@
-from notiblocks.main.nb_config import NBConfig
-from notiblocks.main.nb_handler import NBHandler
-from notiblocks.main.inline_formatting.il_formatter import ILFormatter
+from notiblocks.main._nb_config import NBConfig
+from notiblocks.main._nb_handler import NBHandler
+from notiblocks.main.inline_formatting._il_config import ILConfig
+from notiblocks.main.inline_formatting._il_formatter import ILFormatter
+
+__version__ = "0.0.4"
 
 __all__ = [
     NBConfig,
